refactor: 코드 끝 공백 추가, 불필요한 개행 삭제
