+++ src/main/java/dbdr/domain/careworker/dto/request/CareworkerRequest.java
@@ -36,5 +36,4 @@ public class CareworkerRequest {
     @NotBlank(message = "비밀번호는 필수 항목입니다.")
     private String loginPassword;
 
-
 }
