+++ src/main/java/dbdr/domain/careworker/dto/request/CareworkerUpdateRequest.java
@@ -1,6 +1,5 @@
 package dbdr.domain.careworker.dto.request;
 
-
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
