+++ src/main/java/dbdr/domain/careworker/service/CareworkerService.java
@@ -258,5 +258,4 @@ public class CareworkerService {
         careworkerRepository.save(careworker);
     }
 
-
 }
