+++ src/main/java/dbdr/domain/recipient/controller/RecipientGuardianController.java
@@ -44,4 +44,3 @@ public class RecipientGuardianController {
         return ResponseEntity.ok(ApiUtils.success(recipient));
     }
 }
-
