+++ src/main/java/dbdr/security/model/Role.java
@@ -8,7 +8,4 @@ package dbdr.security.model;
  */
 public enum Role {
     ADMIN,INSTITUTION,GUARDIAN,CAREWORKER;
-
-
 }
-
