+++ src/main/java/dbdr/security/service/DbdrSeucrityService.java
@@ -88,5 +88,4 @@ public class DbdrSeucrityService {
         };
     }
 
-
 }
